Included makedebpkg.sh.
The are two versions of the script: the original (the shell script) and
the Python one.
The Python version has been written to test out whether it was possible
to easily create such a system and to provide a safer work-environment:
indeed, since it doesn't 'source' a bash script (like makedebpkg.sh
does) it is less likely for a malicious PKGBUILD to be executed.
These scripts are not updated since at least 6 years.